requirements.txt
flask
jinja2
sentence-transformers
faiss-cpu
transformers
